Bug 777292 - Make sure that the Components.lastResult is interpreted as unsigned integers; r=bzbarsky
UltraBlame original commit: dbcc29e9fcf66cc5f5501826d6b69e73c8b1cab6

+++ js/xpconnect/src/XPCComponents.cpp
@@ -4721,7 +4721,10 @@ nsXPCComponents::GetProperty(nsIXPConnectWrappedNative *wrapper,
 
     nsresult rv = NS_OK;
     if (doResult) {
-        *vp = JS_NumberValue((double) res);
+        
+        
+        
+        *vp = JS_NumberValue((double)(uint32_t) res);
         rv = NS_SUCCESS_I_DID_SOMETHING;
     }
